front-end/src/test/javascript/spec/app/entities/ApiMonitoramentoBarragens/afetado/afetado.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { AfetadoService } from 'app/entities/ApiMonitoramentoBarragens/afetado/afetado.service';
import { IAfetado, Afetado } from 'app/shared/model/ApiMonitoramentoBarragens/afetado.model';

describe('Service Tests', () => {
  describe('Afetado Service', () => {
    let injector: TestBed;
    let service: AfetadoService;
    let httpMock: HttpTestingController;
    let elemDefault: IAfetado;
    let expectedResult: IAfetado | IAfetado[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(AfetadoService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Afetado(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Afetado', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Afetado()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Afetado', () => {
        const returnedFromService = Object.assign(
          {
            nome: 'BBBBBB',
            email: 'BBBBBB',
            telefonePrincipal: 'BBBBBB',
            telefoneReserva: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Afetado', () => {
        const returnedFromService = Object.assign(
          {
            nome: 'BBBBBB',
            email: 'BBBBBB',
            telefonePrincipal: 'BBBBBB',
            telefoneReserva: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Afetado', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
